macOS changes related to 1LU not being considered the same type as uint64_t

+++ src/duplication_results.cpp
@@ -121,14 +121,16 @@ duplication_results::get_overrepresented(const std::uint64_t n_reads) const
   std::ranges::sort(overrep, std::greater{});
   std::vector<overrep_t> ret;
   for (const auto &[n_obs, seq] : overrep)
-    ret.emplace_back(seq, n_obs, pct(as_frac(n_obs, std::max(1LU, n_reads))),
-                     match_contaminant(seq.string()));
+    ret.emplace_back(
+      seq, n_obs,
+      pct(as_frac(n_obs, std::max(static_cast<std::uint64_t>(1), n_reads))),
+      match_contaminant(seq.string()));
   return ret;
 }
 
 auto
-duplication_results::initialize(const run_mode &mode,
-                                const file_info &info) -> void {
+duplication_results::initialize(const run_mode &mode, const file_info &info)
+  -> void {
   read_skip =
     info.n_reads_est < max_n_reads_total
       ? 0
@@ -160,8 +162,9 @@ get_grade_overrepresented(const std::uint64_t n_reads,
   static constexpr auto label = "overrepresented";
   const auto max_n_obs =
     dr.dups.empty() ? 0LU : std::ranges::max(std::views::values(dr.dups));
-  return grader_set::get_grade(label,
-                               as_frac(max_n_obs, std::max(1LU, n_reads)));
+  return grader_set::get_grade(
+    label,
+    as_frac(max_n_obs, std::max(static_cast<std::uint64_t>(1), n_reads)));
 }
 
 [[nodiscard]] auto
@@ -275,15 +278,16 @@ get_grade_duplication(const dup_summary_t &summary) -> std::string {
 }
 
 [[nodiscard]] auto
-duplication_report(const dup_summary_t &summary,
-                   const file_grades &grades) -> std::string {
+duplication_report(const dup_summary_t &summary, const file_grades &grades)
+  -> std::string {
   static constexpr auto label = "duplication";
   static constexpr auto start_tag = ">>Sequence Duplication Levels\t{}\n"
                                     "#Total Deduplicated Percentage\t{:.6f}\n";
   static constexpr auto header = "#Duplication Level\t"
                                  "Percentage of total\n";
   const auto to_pct = [&](const auto &v) {
-    const auto sum = std::max(1LU, std::reduce(std::cbegin(v), std::cend(v)));
+    const auto sum = std::max(static_cast<std::uint64_t>(1),
+                              std::reduce(std::cbegin(v), std::cend(v)));
     const auto p = [&](const auto d) { return pct(as_frac(d, sum)); };
     return make_bins(bin_breaks, v) | std::views::transform(p) |
            std::ranges::to<std::vector>();
@@ -292,8 +296,9 @@ duplication_report(const dup_summary_t &summary,
   const auto reduce = [](const auto &v) {
     return std::reduce(std::cbegin(v), std::cend(v));
   };
-  const auto frac_dedup = as_frac(reduce(summary.hist_dedup),
-                                  std::max(1LU, reduce(summary.hist_mass)));
+  const auto frac_dedup =
+    as_frac(reduce(summary.hist_dedup),
+            std::max(static_cast<std::uint64_t>(1), reduce(summary.hist_mass)));
   auto r = std::format(start_tag, grades.grade(label), pct(frac_dedup));
   r += header;
   for (const auto [label, mass] :
@@ -336,8 +341,8 @@ overrepresented_html(const std::vector<overrep_t> &overrep,
 }
 
 [[nodiscard]] auto
-duplication_html(const dup_summary_t &summary,
-                 const file_grades &grades) -> std::string {
+duplication_html(const dup_summary_t &summary, const file_grades &grades)
+  -> std::string {
   static constexpr auto label = "duplication";
   static constexpr auto plot_format = R"(<div id="duplication_plot"></div>
 <script>Plotly.newPlot("duplication_plot",
@@ -372,7 +377,8 @@ yaxis: {{title: "% of sequences"}},
   static constexpr auto x_text =  // has one extra element at "0"
     std::ranges::subrange(std::cbegin(bin_labels), std::cend(bin_labels));
   const auto to_pct = [&](const auto &v) {
-    const auto sum = std::max(1LU, std::reduce(std::cbegin(v), std::cend(v)));
+    const auto sum = std::max(static_cast<std::uint64_t>(1),
+                              std::reduce(std::cbegin(v), std::cend(v)));
     const auto p = [&](const auto d) { return pct(as_frac(d, sum)); };
     return make_bins(bin_breaks, v) | std::views::transform(p) |
            std::ranges::to<std::vector>();
